backend/requirements.txt
asgiref==3.7.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.2
django-annoying==0.10.6
django-cors-headers==4.1.0
django-extensions==3.2.3
django-filter==23.2
django-jsonstore==0.5.1
django-soft-delete==0.9.21
djangorestframework==3.14.0
drf-yasg==1.21.5
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.4.3
markdown-include==0.8.1
MarkupSafe==2.1.3
openapi-codec==1.3.2
packaging==23.1
pycparser==2.21
PyNaCl==1.5.0
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
requests==2.31.0
ruamel.yaml==0.17.31
ruamel.yaml.clib==0.2.7
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==2.0.3
